requirements.txt
flask
python-dotenv
requests
gunicorn
openai==1.12.0
httpx==0.27.2
